# swagger-spec.json
{"openapi":"3.0.0","paths":{"/aggregator/logs":{"get":{"description":"Get contract logs from the database","operationId":"AggregatorContorller_getLogs","parameters":[{"name":"startBlock","required":false,"in":"query","description":"Block number from which the logs are starting","schema":{"type":"number"}},{"name":"endBlock","required":false,"in":"query","description":"Block number from which the logs are ending","schema":{"type":"number"}},{"name":"startDate","required":false,"in":"query","description":"Start date of the logs. Used timestamp in milliseconds","schema":{"type":"string"}},{"name":"endDate","required":false,"in":"query","description":"End date of the logs. Used timestamp in milliseconds","schema":{"type":"string"}},{"name":"from","required":false,"in":"query","description":"Address of the sender, that will be used as a filter","schema":{"type":"string"}},{"name":"to","required":false,"in":"query","description":"Address of the receiver, that will be used as a filter","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","description":"Limit of the logs that will be returned","schema":{"minimum":0,"maximum":10000,"type":"number"}},{"name":"offset","required":false,"in":"query","description":"Offset of the logs that will be returned","schema":{"minimum":0,"type":"number"}},{"name":"order","required":false,"in":"query","description":"Order of the logs that will be returned","schema":{"default":"DESC","type":"string","enum":["ASC","DESC"]}}],"responses":{"200":{"description":"Logs were successfully fetched","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetLogsQueryResponseDTO"}},"example":[{"transactionHash":"0xaefe609781e88d91e5b6e0ee0c4295b385614f3d1edca3e5474fdf6b0f82509f","blockNumber":20391007,"executedDate":null,"uploaded":"2025-02-20T18:19:59.114Z","from":"0x2C88A3B97263D703fD6D2434e14fA3a51f3Ce39c","to":"0xf4003F4efBE8691B60249E6afbD307aBE7758adb","value":"1921"}]}}}},"summary":"Get logs","tags":["Aggregator"]}},"/aggregator/total":{"get":{"description":"Get total transactions and values for all indexed period. Also possible to select specific period by date in timesamp or block number","operationId":"AggregatorContorller_getTotal","parameters":[{"name":"startBlock","required":false,"in":"query","description":"Block number from which the logs are starting","schema":{"type":"number"}},{"name":"endBlock","required":false,"in":"query","description":"Block number from which the logs are ending","schema":{"type":"number"}},{"name":"startDate","required":false,"in":"query","description":"Start date of the logs. Used timestamp in milliseconds","schema":{"type":"string"}},{"name":"endDate","required":false,"in":"query","description":"End date of the logs. Used timestamp in milliseconds","schema":{"type":"string"}}],"responses":{"200":{"description":"Object with total values for selected or all period","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TotalResponseDTO"},"example":{"firstTransaction":"1970-01-01T00:00:00.000Z","lastTransaction":"2022-09-22T21:20:34.000Z","firstBlock":7390439,"lastBlock":21048005,"totalTransactionsCount":"5476041","totalValue":"98206728778933391","minTransactionValue":"0","maxTransactionValue":"193056413000000"}}}}},"summary":"Get Total","tags":["Aggregator"]}},"/aggregator/top-accounts":{"get":{"description":"Get top accounts by total value of transactions. Also possible to select specific period by date in timesamp or block number","operationId":"AggregatorContorller_getTopAccounts","parameters":[{"name":"startBlock","required":false,"in":"query","description":"Block number from which the logs are starting","schema":{"type":"number"}},{"name":"endBlock","required":false,"in":"query","description":"Block number from which the logs are ending","schema":{"type":"number"}},{"name":"startDate","required":false,"in":"query","description":"Start date of the logs. Used timestamp in milliseconds","schema":{"type":"string"}},{"name":"endDate","required":false,"in":"query","description":"End date of the logs. Used timestamp in milliseconds","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","description":"Limit of the results","schema":{"minimum":0,"maximum":10000,"type":"number"}},{"name":"offset","required":false,"in":"query","description":"Offset of the results","schema":{"minimum":0,"type":"number"}},{"name":"order","required":false,"in":"query","description":"Order direction of the results","schema":{"default":"DESC","type":"string","enum":["ASC","DESC"]}},{"name":"orderBy","required":false,"in":"query","description":"Order by which the results are ordered","schema":{"default":"totalValue","type":"string","enum":["totalValue"]}},{"description":"Block number from which the logs are starting","required":false,"name":"startBlock","in":"query","schema":{"type":"number"}},{"description":"Block number from which the logs are ending","required":false,"name":"endBlock","in":"query","schema":{"type":"number"}},{"description":"Start date of the logs. Used timestamp in milliseconds","required":false,"name":"startDate","in":"query","schema":{"type":"string"}},{"description":"End date of the logs. Used timestamp in milliseconds","required":false,"name":"endDate","in":"query","schema":{"type":"string"}},{"description":"Limit of the results","required":false,"name":"limit","in":"query","schema":{"minimum":0,"maximum":10000,"type":"number"}},{"description":"Offset of the results","required":false,"name":"offset","in":"query","schema":{"minimum":0,"type":"number"}},{"description":"Order direction of the results","required":false,"name":"order","in":"query","schema":{"default":"DESC","type":"string"}},{"description":"Order by which the results are ordered","required":false,"name":"orderBy","in":"query","schema":{"default":"totalValue","type":"string"}}],"responses":{"200":{"description":"List of top accounts","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TopAccountsResponseDTO"}},"example":[{"address":"0x625E7708f30cA75bfd92586e17077590C60eb4cD","totalValue":"15609044266552454","totalTransactions":"60059"},{"address":"0xBF14DB80D9275FB721383a77C00Ae180fc40ae98","totalValue":"6905980867490000","totalTransactions":"13822"},{"address":"0x279f8940ca2a44C35ca3eDf7d28945254d0F0aE6","totalValue":"5103012195406319","totalTransactions":"28575"}]}}}},"summary":"Get Top Accounts List","tags":["Aggregator"]}},"/analyzer/logs":{"get":{"description":"Directly from blockchain. This endpoint able to get max 1000 logs","operationId":"AnalyzerController_getLogs","parameters":[{"name":"startBlock","required":true,"in":"query","description":"Block number from which the logs are starting","schema":{"type":"number"}},{"name":"endBlock","required":false,"in":"query","description":"Block number from which the logs are ending","schema":{"type":"number"}},{"name":"from","required":false,"in":"query","description":"Start date of the logs. Used timestamp in milliseconds","schema":{"type":"string"}},{"name":"to","required":false,"in":"query","description":"End date of the logs. Used timestamp in milliseconds","schema":{"type":"string"}},{"name":"topicName","required":false,"in":"query","description":"Topic name of the logs","schema":{"type":"string"}}],"responses":{"200":{"description":"Logs successfully fetched","content":{"application/json":{"example":[{"transactionHash":"0x2b326a283dbbfe4cff4ed4609b76cec49308983cc572ff730ed55df875fc7c29","blockNumber":7390439,"logIndex":1,"from":"0x0000000000000000000000000000000000000000","to":"0x4C90c8a368F7d1F4a57906139bAc23C72f38F4cf","value":"200000"},{"transactionHash":"0x4139234f1f9618d6ef07cbc5812a058004a0814873a01a81889771c592009390","blockNumber":7390447,"logIndex":1,"from":"0x0000000000000000000000000000000000000000","to":"0xB5dc6671e7cdAaA2386536295BF63A221b67311a","value":"200000"},{"transactionHash":"0xd98180a35a0b5b493cdec899adfae243722f14dba561504614189757e97a9009","blockNumber":7390466,"logIndex":7,"from":"0x4C90c8a368F7d1F4a57906139bAc23C72f38F4cf","to":"0xcb9968Cb0d6612e1167e445774997C63a0792dbF","value":"100000"}]}}}},"summary":"Get logs","tags":["Analyzer"]}}},"info":{"title":" API","description":"The API description","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"api/v1"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"GetLogsQueryResponseDTO":{"type":"object","properties":{"transactionHash":{"type":"string","description":"Transaction hash"},"blockNumber":{"type":"number","description":"Block number"},"logIndex":{"type":"number","description":"Log index"},"from":{"type":"string","description":"From address"},"to":{"type":"string","description":"To address"},"value":{"type":"string","description":"Value of the transaction"},"executedDate":{"type":"string","description":"Date when the transaction was executed"},"uploaded":{"type":"string","description":"Date when the log was uploaded"}},"required":["transactionHash","blockNumber","logIndex","from","to","value","executedDate","uploaded"]},"TotalResponseDTO":{"type":"object","properties":{"firstTransaction":{"type":"string","description":"First transaction hash"},"lastTransaction":{"type":"string","description":"Last transaction hash"},"firstBlock":{"type":"number","description":"First block number"},"lastBlock":{"type":"number","description":"Last block number"},"totalTransactionsCount":{"type":"number","description":"Total number of transactions"},"totalValue":{"type":"number","description":"Total value of transactions"},"minTransactionValue":{"type":"number","description":"Minimum transaction value"},"maxTransactionValue":{"type":"number","description":"Maximum transaction value"}},"required":["firstTransaction","lastTransaction","firstBlock","lastBlock","totalTransactionsCount","totalValue","minTransactionValue","maxTransactionValue"]},"TopAccountsResponseDTO":{"type":"object","properties":{"address":{"type":"string","description":"Address of the account"},"totalValue":{"type":"string","description":"Total value of the account"},"totalTransactions":{"type":"number","description":"Total number of transactions"}},"required":["address","totalValue","totalTransactions"]}}}}